change functions, that works with DB using Mongoose
